new helper: __alloc_fd()
Essentially, alloc_fd() in a files_struct we own a reference to.
Most of the time wanting to use it is a sign of lousy API
design (such as android/binder).  It's *not* a general-purpose
interface; better that than open-coding its guts, but again,
playing with other process' descriptor table is a sign of bad
design.

Change-Id: I3d1b83cdd19611e6908aaeefa69a579c1d0c6674
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>