vfs: Allow filesystems to access their private mount data
Now we pass the vfsmount when mounting and remounting.
This allows the filesystem to actually set up the mount
specific data, although we can't quite do anything with
it yet. show_options is expanded to include data that
lives with the mount.

To avoid changing existing filesystems, these have
been added as new vfs functions.

Change-Id: I996dbef3451b9bcfed5e44f0ac19368921c9cae1
Signed-off-by: Daniel Rosenberg <drosen@google.com>