ANDROID: sdcardfs: Don't bother deleting freelist
There is no point deleting entries from dlist, as
that is a temporary list on the stack from which
contains only entries that are being deleted.

Not all code paths set up dlist, so those that
don't were performing invalid accesses in
hash_del_rcu. As an additional means to prevent
any other issue, we null out the list entries when
we allocate from the cache.

Change-Id: Ibcc37aaf8d54e1a345c90807955760df1a9543cd
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 35666680